+++ shardingsphere-proxy/shardingsphere-proxy-bootstrap/src/test/java/org/apache/shardingsphere/proxy/database/DatabaseServerInfoTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.db;
+package org.apache.shardingsphere.proxy.database;
 
 import org.apache.shardingsphere.infra.exception.ShardingSphereException;
 import org.junit.Test;
